move list layout to plugins
