+++ api/src/main/java/edu/wpi/first/shuffleboard/api/widget/ParametrizedController.java
@@ -17,5 +17,4 @@ public @interface ParametrizedController {
    * The path to the FXML file for the widget with this annotation.
    */
   String value();
-
 }
