+++ app/src/main/java/edu/wpi/first/shuffleboard/app/WidgetPaneController.java
@@ -17,8 +17,8 @@ import edu.wpi.first.shuffleboard.app.components.TileLayout;
 import edu.wpi.first.shuffleboard.app.components.WidgetPane;
 import edu.wpi.first.shuffleboard.app.components.WidgetTile;
 import edu.wpi.first.shuffleboard.app.dnd.TileDragResizer;
-import edu.wpi.first.shuffleboard.app.widget.Layout;
-import edu.wpi.first.shuffleboard.app.widget.ListLayout;
+import edu.wpi.first.shuffleboard.api.components.Layout;
+import edu.wpi.first.shuffleboard.plugin.base.layout.ListLayout;
 
 import org.fxmisc.easybind.EasyBind;
