+++ app/src/main/java/edu/wpi/first/shuffleboard/app/components/LayoutTile.java
@@ -1,6 +1,6 @@
 package edu.wpi.first.shuffleboard.app.components;
 
-import edu.wpi.first.shuffleboard.app.widget.Layout;
+import edu.wpi.first.shuffleboard.api.components.Layout;
 
 public class LayoutTile extends Tile<Layout> {
   public LayoutTile(Layout container) {
