+++ plugins/base/src/main/java/edu/wpi/first/shuffleboard/plugin/base/layout/ListLayout.java
@@ -1,5 +1,6 @@
-package edu.wpi.first.shuffleboard.app.widget;
+package edu.wpi.first.shuffleboard.plugin.base.layout;
 
+import edu.wpi.first.shuffleboard.api.components.Layout;
 import edu.wpi.first.shuffleboard.api.widget.ParametrizedController;
 import edu.wpi.first.shuffleboard.api.widget.Component;
 import edu.wpi.first.shuffleboard.api.components.EditableLabel;
