+++ plugins/base/src/main/resources/edu/wpi/first/shuffleboard/plugin/base/layout/ListLayout.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.control.ScrollPane?>
 <StackPane xmlns="http://javafx.com/javafx"
            xmlns:fx="http://javafx.com/fxml"
-           fx:controller="edu.wpi.first.shuffleboard.app.widget.ListLayout"
+           fx:controller="edu.wpi.first.shuffleboard.plugin.base.layout.ListLayout"
            fx:id="root"
            styleClass="layout--root"
            prefHeight="128" prefWidth="128">
